requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-browser-reload==1.12.1
django-crispy-forms==2.1
django-dsfr==0.18.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
idna==3.6
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
